Coloca a funcao templateUpgrade e as fontes na config e importa elas onde e necessario (Coesao)
