+++ save.json
@@ -1,1 +1,1 @@
-{"Dinheiro": 788, "Soma": 11, "Gemas": 0, "Up1": 10, "Up2": 6, "Up3": 0, "Up4": 0, "Up5": 0, "Up6": 0, "Auto": 60, "Missao": 6, "Clicks": 90, "Acumulado": 942, "AcumuladoAuto": 1790}
+{"Dinheiro": 512, "Soma": 12, "Gemas": 0, "Up1": 11, "Up2": 6, "Up3": 0, "Up4": 0, "Up5": 0, "Up6": 0, "Auto": 60, "Missao": 6, "Clicks": 90, "Acumulado": 942, "AcumuladoAuto": 2090}
